Change progress bar logic to not compare all videos beforehand
The old progress bar was based on number of matching videos.
This was very accurate, but all videos had to be compared with each other first.
Furthermore, this had to be done after every delete/move operation.
All this meant there was a noticeable delay between videos.

The new progress bar is based on the current position out of all possible comparisons.
Not a single video needs to be compared, making it update very fast.

There is a downside: Progress bar is no longer incremented in nice, equal steps,
but skips ahead randomly until the next matching video.